Refactor: useUserQuery를 하위 컴포넌트로 이동
- Collections에서 Collection으로 이동
- 로직과 연관된 Pagination, EmptyMessgae 컴포넌트 함께 이동
