+++ src/components/myPage/Collections/Collection.jsx
@@ -1,15 +1,26 @@
 import { useRef, useState } from 'react';
-import { Accordion } from '@mantine/core';
+import { Accordion, Pagination } from '@mantine/core';
 import { useRecoilValue } from 'recoil';
-import { useCollectionQueries } from '../../../hooks/queries';
-import { useSelectedItem } from '../../../hooks';
-import { sideNavState } from '../../../recoil/atom';
-import { Item } from '.';
+import { useMediaQuery } from '@mantine/hooks';
+import { useCollectionQueries, useUserQuery } from '../../../hooks/queries';
+import { useSelectedItem, usePagination } from '../../../hooks';
+import { sideNavState, categoryState } from '../../../recoil/atom';
+import { EmptyMessage, Item } from '.';
 
-const Collection = ({ collection, setIsItemSelected, setImgSrc, page }) => {
+const Collection = ({ setIsItemSelected, setImgSrc, page }) => {
+  const smallScreen = useMediaQuery('(max-width: 48rem)');
+
+  const category = useRecoilValue(categoryState);
   const isNavOpened = useRecoilValue(sideNavState);
 
-  const collectionQueries = useCollectionQueries(collection);
+  const { userInfo = [] } = useUserQuery({
+    select: userInfo => userInfo[`${category}_list`],
+    refetchOnWindowFocus: false,
+  });
+
+  const { activePage, setActivePage, total, collection } = usePagination(userInfo);
+
+  const collectionQueries = useCollectionQueries(collection, { enable: !!collection });
 
   const [selectedItem, setSelectedItem] = useState(null);
 
@@ -42,11 +53,29 @@ const Collection = ({ collection, setIsItemSelected, setImgSrc, page }) => {
 
   return (
     <>
-      <Accordion variant="separated" w="100%" onChange={selectItem} value={selectedItem}>
-        {collectionList?.map(item => (
-          <Item key={item.id} item={item} setSelectedItem={setSelectedItem} setIsItemSelected={setIsItemSelected} />
-        ))}
-      </Accordion>
+      {collectionList.length === 0 ? (
+        <EmptyMessage category={category} />
+      ) : (
+        <>
+          <Accordion variant="separated" w="100%" onChange={selectItem} value={selectedItem}>
+            {collectionList?.map(item => (
+              <Item key={item.id} item={item} setSelectedItem={setSelectedItem} setIsItemSelected={setIsItemSelected} />
+            ))}
+          </Accordion>
+
+          <Pagination
+            value={activePage}
+            onChange={setActivePage}
+            total={total}
+            siblings={2}
+            withEdges
+            align="center"
+            position="center"
+            size={smallScreen ? 'xs' : 'sm'}
+            m={smallScreen ? 'xs' : 'sm'}
+          />
+        </>
+      )}
     </>
   );
 };
